+++ src/test/java/com/moa/moa_server/unit/vote/service/VoteUpdateTest.java
@@ -11,7 +11,7 @@ import com.moa.moa_server.domain.vote.dto.request.VoteUpdateRequest;
 import com.moa.moa_server.domain.vote.dto.response.VoteUpdateResponse;
 import com.moa.moa_server.domain.vote.entity.Vote;
 import com.moa.moa_server.domain.vote.repository.VoteRepository;
-import com.moa.moa_server.domain.vote.service.VoteService;
+import com.moa.moa_server.domain.vote.service.VoteCommandService;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
@@ -30,7 +30,7 @@ public class VoteUpdateTest {
   @Mock VoteRepository voteRepository;
   @Mock ImageService imageService;
 
-  @InjectMocks VoteService voteService;
+  @InjectMocks VoteCommandService voteCommandService;
 
   @Nested
   class 성공_케이스 {
@@ -54,7 +54,7 @@ public class VoteUpdateTest {
           .thenReturn(new ImageProcessResult("vote/abc.jpg", "파일이름.jpeg"));
 
       // when
-      VoteUpdateResponse resp = voteService.updateVote(userId, voteId, req);
+      VoteUpdateResponse resp = voteCommandService.updateVote(userId, voteId, req);
 
       // then
       assertThat(resp.voteId()).isEqualTo(voteId); // 서비스 응답 값 검사
@@ -94,7 +94,7 @@ public class VoteUpdateTest {
           .thenReturn(new ImageProcessResult(oldImageUrl, oldImageName));
 
       // when
-      VoteUpdateResponse resp = voteService.updateVote(userId, voteId, req);
+      VoteUpdateResponse resp = voteCommandService.updateVote(userId, voteId, req);
 
       // then
       assertThat(resp.voteId()).isEqualTo(voteId);
@@ -131,7 +131,7 @@ public class VoteUpdateTest {
           .thenReturn(new ImageProcessResult("", ""));
 
       // when
-      VoteUpdateResponse resp = voteService.updateVote(userId, voteId, req);
+      VoteUpdateResponse resp = voteCommandService.updateVote(userId, voteId, req);
 
       // then
       assertThat(resp.voteId()).isEqualTo(voteId);
